+++ apps/docs/src/examples/index.ts
@@ -10,7 +10,27 @@
   };
   
   export const Demos: Record<string, Demo> = {
-    "badge/alert": {
+    "avatar/initials": {
+	  component: lazy(() => import("~/examples/avatar/initials.tsx")),
+	  code: `import { Avatar } from "@pixelshades/ui/components"
+
+export function Example() {
+	return <Avatar initials="ma" alt="@makisuo__" />
+}
+`,
+  },
+  "avatar/preview": {
+	  component: lazy(() => import("~/examples/avatar/preview.tsx")),
+	  code: `import { Avatar } from "@pixelshades/ui/components"
+
+export function Example() {
+	return (
+		<Avatar src="https://pbs.twimg.com/profile_images/1271783863614873600/DozTdIj7_400x400.jpg" alt="@makisuo__" />
+	)
+}
+`,
+  },
+  "badge/alert": {
 	  component: lazy(() => import("~/examples/badge/alert.tsx")),
 	  code: `import { Badge } from "@pixelshades/ui/components"
 import { ArrowRightIcon, SmileIcon } from "@pixelshades/ui/icons"
@@ -886,8 +906,10 @@ export function Example() {
   },
   "drawer/directions": {
 	  component: lazy(() => import("~/examples/drawer/directions.tsx")),
-	  code: `import { Button, Drawer, type DrawerPlacement, PLACEMENTS, Select } from "@pixelshades/ui/components"
+	  code: `import { Button, Drawer, type DrawerPlacement, Select } from "@pixelshades/ui/components"
 import React from "react"
+
+export const PLACEMENTS = ["bottom", "left", "right", "top"] satisfies DrawerPlacement[]
 
 export function Example() {
 	const options = PLACEMENTS.map((d) => ({ name: d }))
@@ -954,8 +976,9 @@ export function Example() {
   },
   "drawer/placement": {
 	  component: lazy(() => import("~/examples/drawer/placement.tsx")),
-	  code: `import { Button, Drawer, type DrawerPlacement, PLACEMENTS, Select } from "@pixelshades/ui/components"
+	  code: `import { Button, Drawer, type DrawerPlacement, Select } from "@pixelshades/ui/components"
 import React from "react"
+import { PLACEMENTS } from "~/examples/drawer/directions.tsx"
 
 export function Example() {
 	const options = PLACEMENTS.map((d) => ({ name: d }))
@@ -1029,7 +1052,7 @@ export function Example() {
 	DateField,
 	DatePicker,
 	Form,
-	Radio,
+	RadioGroup,
 	Switch,
 	TextField,
 	TimeField,
@@ -1073,17 +1096,17 @@ export function Example() {
 
 			<Switch label="Switch" description="Description" tooltip="Tooltip" helperText="Helper text" />
 
-			<Radio.Group
+			<RadioGroup
 				label="Radio Group"
 				description="Description"
 				tooltip="Tooltip"
 				helperText="Helper Text"
 				isRequired
 			>
-				<Radio value="push">Push</Radio>
-				<Radio value="email">Email</Radio>
-				<Radio value="text">Text</Radio>
-			</Radio.Group>
+				<RadioGroup.Item value="push">Push</RadioGroup.Item>
+				<RadioGroup.Item value="email">Email</RadioGroup.Item>
+				<RadioGroup.Item value="text">Text</RadioGroup.Item>
+			</RadioGroup>
 			<div className="flex w-full flex-row gap-md">
 				<Button className="w-full" type="submit">
 					Submit
@@ -1631,18 +1654,54 @@ export function Example() {
 }
 `,
   },
-  "radio-group/preview": {
-	  component: lazy(() => import("~/examples/radio-group/preview.tsx")),
-	  code: `import { Label, Radio } from "@pixelshades/ui/components"
+  "radio-group/card": {
+	  component: lazy(() => import("~/examples/radio-group/card.tsx")),
+	  code: `import { RadioGroup } from "@pixelshades/ui/components"
 
 export function Example() {
 	return (
-		<Radio.Group defaultValue="dog">
+		<RadioGroup defaultValue="dog" orientation="horizontal" label="Favorite pet" description="Description">
+			<RadioGroup.Card value="dog">Dog</RadioGroup.Card>
+			<RadioGroup.Card value="cat">Cat</RadioGroup.Card>
+			<RadioGroup.Card value="dragon">Dragon</RadioGroup.Card>
+		</RadioGroup>
+	)
+}
+`,
+  },
+  "radio-group/horizontal": {
+	  component: lazy(() => import("~/examples/radio-group/horizontal.tsx")),
+	  code: `import { RadioGroup } from "@pixelshades/ui/components"
+
+export function Example() {
+	return (
+		<RadioGroup
+			defaultValue="dog"
+			orientation="horizontal"
+			label="Favorite pet"
+			description="Description"
+			helperText="WOW"
+		>
+			<RadioGroup.Item value="dog">Dog</RadioGroup.Item>
+			<RadioGroup.Item value="cat">Cat</RadioGroup.Item>
+			<RadioGroup.Item value="dragon">Dragon</RadioGroup.Item>
+		</RadioGroup>
+	)
+}
+`,
+  },
+  "radio-group/preview": {
+	  component: lazy(() => import("~/examples/radio-group/preview.tsx")),
+	  code: `import { Label, RadioGroup } from "@pixelshades/ui/components"
+
+export function Example() {
+	return (
+		<RadioGroup defaultValue="dog">
 			<Label>Favorite pet</Label>
-			<Radio value="dog">Dog</Radio>
-			<Radio value="cat">Cat</Radio>
-			<Radio value="dragon">Dragon</Radio>
-		</Radio.Group>
+			<RadioGroup.Item value="dog">Dog</RadioGroup.Item>
+			<RadioGroup.Item value="cat">Cat</RadioGroup.Item>
+			<RadioGroup.Item value="dragon">Dragon</RadioGroup.Item>
+		</RadioGroup>
 	)
 }
 `,