+++ packages/ui/src/components/ui/drawer/drawer.tsx
@@ -5,14 +5,11 @@ import type { VariantProps } from "tailwind-variants"
 import { DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogTrigger } from "../dialog"
 import { Modal, type ModalOverlayProps } from "../modal"
 
-export const PLACEMENTS = ["bottom", "left", "right", "top"] as const
-export type DrawerPlacement = (typeof PLACEMENTS)[number]
-
-export type Placement = VariantProps<typeof drawerVariants>["placement"]
+export type DrawerPlacement = VariantProps<typeof drawerVariants>["placement"]
 
 export interface DrawerProps extends ModalOverlayProps, VariantProps<typeof drawerVariants> {
 	/** The placement of the drawer. */
-	placement?: Placement
+	placement?: DrawerPlacement
 }
 
 const DrawerRoot = ({ placement, children, className, ...props }: DrawerProps) => {